+++ src/rinoh/reference.py
@@ -18,6 +18,7 @@ from .layout import ReflowRequired, ContainerOverflow
 from .number import NumberStyle, Label, format_number
 from .paragraph import Paragraph, ParagraphStyle, ParagraphBase
 from .strings import StringCollection, StringField
+from .style import HasClass, HasClasses
 from .text import (SingleStyledTextBase, MixedStyledTextBase, TextStyle,
                    StyledText, SingleStyledText, MixedStyledText)
 from .util import NotImplementedAttribute
@@ -55,6 +56,10 @@ class ReferenceBase(MixedStyledTextBase):
         if self.style is not None:
             result += ' ({})'.format(self.style)
         return result
+
+    def copy(self, parent=None):
+        return type(self)(self.type, self.link, self.quiet,
+                          style=self.style, parent=parent)
 
     def target_id(self, document):
         raise NotImplementedError
@@ -166,8 +171,28 @@ class ReferencingParagraph(ParagraphBase):
         self.target_id_or_flowable = target_id_or_flowable
 
     def text(self, container):
-        return MixedStyledText(self.get_style('text', container), parent=self)
+        return self.get_style('text', container).copy(parent=self)
 
+    def _target_id(self, document):
+        target_id_or_flowable = self.target_id_or_flowable
+        return (target_id_or_flowable.get_id(document)
+                if isinstance(target_id_or_flowable, Flowable)
+                else target_id_or_flowable)
+
+    def _target_flowable(self, document):
+        return document.elements[self._target_id(document)]
+
+    def target_style(self, document):
+        """Filter selection on the ``style`` attribute of the target flowable"""
+        return self._target_flowable(document).style
+
+    def target_has_class(self, document):
+        """Filter selection on a class of the target flowable"""
+        return HasClass(self._target_flowable(document))
+
+    def target_has_classes(self, document):
+        """Filter selection on a set of classes of the target flowable"""
+        return HasClasses(self._target_flowable(document))
 
 
 class Note(LabeledFlowable):