+++ src/rinoh/style.py
@@ -149,7 +149,7 @@ class Selector(object):
     def flatten(self, stylesheet):
         raise NotImplementedError
 
-    def match(self, styled, stylesheet):
+    def match(self, styled, stylesheet, document):
         raise NotImplementedError
 
 
@@ -179,8 +179,8 @@ class SelectorWithPriority(Selector):
         flattened_selector = self.selector.flatten(stylesheet)
         return flattened_selector.pri(self.priority)
 
-    def match(self, styled, stylesheet):
-        score = self.selector.match(styled, stylesheet)
+    def match(self, styled, stylesheet, document):
+        score = self.selector.match(styled, stylesheet, document)
         if score:
             score = Specificity(self.priority, 0, 0, 0, 0) + score
         return score
@@ -225,9 +225,9 @@ class SelectorByName(SingleSelector):
         selector = matcher.by_name[self.name]
         return selector.get_style_name(matcher)
 
-    def match(self, styled, stylesheet):
+    def match(self, styled, stylesheet, document):
         selector = stylesheet.get_selector(self.name)
-        return selector.match(styled, stylesheet)
+        return selector.match(styled, stylesheet, document)
 
 
 class ClassSelectorBase(SingleSelector):
@@ -245,7 +245,7 @@ class ClassSelectorBase(SingleSelector):
     def get_style_name(self, matcher):
         return self.style_name
 
-    def match(self, styled, document):
+    def match(self, styled, stylesheet, document):
         if not isinstance(styled, self.cls):
             return None
         class_match = 2 if type(styled) == self.cls else 1
@@ -258,7 +258,12 @@ class ClassSelectorBase(SingleSelector):
 
         attributes_match = 0
         for attr, value in self.attributes.items():
-            if not hasattr(styled, attr) or getattr(styled, attr) != value:
+            if not hasattr(styled, attr):
+                return None
+            attr = getattr(styled, attr)
+            if callable(attr):
+                attr = attr(document)
+            if attr != value:
                 return None
             attributes_match += 1
 
@@ -295,7 +300,7 @@ class ContextSelector(Selector):
     def get_style_name(self, matcher):
         return self.selectors[-1].get_style_name(matcher)
 
-    def match(self, styled, stylesheet):
+    def match(self, styled, stylesheet, document):
         def styled_and_parents(element):
             while element is not None:
                 yield element
@@ -310,13 +315,13 @@ class ContextSelector(Selector):
                 element = next(elements)                # NoMoreParentElement
                 if isinstance(selector, EllipsisSelector):
                     selector = next(selectors)          # StopIteration
-                    while not selector.match(element, stylesheet):
+                    while not selector.match(element, stylesheet, document):
                         element = next(elements)        # NoMoreParentElement
             except NoMoreParentElement:
                 return None
             except StopIteration:
                 break
-            score = selector.match(element, stylesheet)
+            score = selector.match(element, stylesheet, document)
             if not score:
                 return None
             total_score += score
@@ -454,12 +459,11 @@ class Styled(DocumentElement, Configurable, metaclass=StyledMeta):
             return 0
 
     def configuration_name(self, document):
-        try:
-            # raise AttributeError
+        try:    # TODO: make document hashable so @cached can be used?
             return self._style_name
         except AttributeError:
             ruleset = self.configuration_class.get_ruleset(document)
-            self._style_name = ruleset.find_style(self)
+            self._style_name = ruleset.find_style(self, document)
             return self._style_name
 
     def fallback_to_parent(self, attribute):
@@ -476,10 +480,12 @@ class Styled(DocumentElement, Configurable, metaclass=StyledMeta):
 
     @property
     def has_class(self):
+        """Filter selection on a class of this :class:`Styled`"""
         return HasClass(self)
 
     @property
     def has_classes(self):
+        """Filter selection on a set of classes of this :class:`Styled`"""
         return HasClasses(self)
 
     def before_placing(self, container):
@@ -559,7 +565,7 @@ class StyledMatcher(dict):
         for name, selector in dict(iterable or (), **kwargs).items():
             self[name] = selector
 
-    def match(self, styled, stylesheet):
+    def match(self, styled, stylesheet, document):
         for cls in type(styled).__mro__:
             if cls not in self:
                 continue
@@ -567,7 +573,7 @@ class StyledMatcher(dict):
             for style in set((style_str, None)):
                 for name, selector in self[cls].get(style, {}).items():
                     selector = selector.flatten(stylesheet)
-                    specificity = selector.match(styled, stylesheet)
+                    specificity = selector.match(styled, stylesheet, document)
                     if specificity:
                         yield Match(name, specificity)
 
@@ -680,14 +686,14 @@ class StyleSheet(RuleSet, Resource):
             else:
                 raise KeyError("No selector found for style '{}'".format(name))
 
-    def find_matches(self, styled):
-        for match in self.matcher.match(styled, self):
+    def find_matches(self, styled, document):
+        for match in self.matcher.match(styled, self, document):
             yield match
         if self.base is not None:
-            yield from self.base.find_matches(styled)
+            yield from self.base.find_matches(styled, document)
 
-    def find_style(self, styled):
-        matches = sorted(self.find_matches(styled),
+    def find_style(self, styled, document):
+        matches = sorted(self.find_matches(styled, document),
                          key=attrgetter('specificity'), reverse=True)
         last_match = Match(None, ZERO_SPECIFICITY)
         for match in matches:
@@ -831,25 +837,34 @@ def parse_class_selector(chars):
 
 def parse_selector_args(chars):
     args, kwargs = [], {}
-    assert next(chars) == '('
+    if next(chars) != '(':
+        raise StyleParseError('Expecting an opening brace')
     eat_whitespace(chars)
     while chars.peek() not in (None, ')'):
-        argument = parse_value(chars)
-        if argument is not None:
-            assert not kwargs
-            args.append(argument)
-        else:
-            keyword = parse_keyword(chars)
+        argument, unknown_keyword = parse_value(chars)
+        eat_whitespace(chars)
+        if chars.peek() == '=':
+            next(chars)
+            keyword = argument
+            if not unknown_keyword:
+                raise StyleParseError("'{}' is not a valid keyword argument"
+                                      .format(keyword))
             eat_whitespace(chars)
-            kwargs[keyword] = parse_value(chars)
+            argument, unknown_keyword = parse_value(chars)
+            kwargs[keyword] = argument
+        elif kwargs:
+            raise StyleParseError('Non-keyword argument cannot follow a '
+                                  'keyword argument')
+        else:
+            args.append(argument)
+        if unknown_keyword:
+            raise StyleParseError("Unknown keyword '{}'".format(argument))
         eat_whitespace(chars)
-        char = next(chars)
-        if char == ')':
-            break
-        assert char == ','
-        eat_whitespace(chars)
-    else:
-        assert next(chars) == ')'
+        if chars.peek() == ',':
+            next(chars)
+            eat_whitespace(chars)
+    if next(chars) != ')':
+        raise StyleParseError('Expecting a closing brace')
     return args, kwargs
 
 
@@ -889,24 +904,16 @@ class CharIterator(str):
             return None
 
 
-def parse_keyword(chars):
-    keyword = chars.match(string.ascii_letters + string.digits + '_')
-    eat_whitespace(chars)
-    if chars.peek() != '=':
-        raise StyleParseError('Expecting an equals sign to follow a keyword')
-    next(chars)
-    return keyword
-
-
 def parse_value(chars):
+    unknown_keyword = False
     first_char = chars.peek()
     if first_char in ("'", '"'):
         value = parse_string(chars)
     elif first_char.isnumeric() or first_char in '+-':
         value = parse_number(chars)
     else:
-        value = None
-    return value
+        value, unknown_keyword = parse_keyword(chars)
+    return value, unknown_keyword
 
 
 def parse_string(chars):
@@ -930,6 +937,17 @@ def parse_string(chars):
 
 def parse_number(chars):
     return literal_eval(chars.match('0123456789.e+-'))
+
+
+def parse_keyword(chars):
+    keyword = chars.match(string.ascii_letters + string.digits + '_')
+    try:
+        return KEYWORDS[keyword.lower()], False
+    except KeyError:
+        return keyword, True
+
+
+KEYWORDS = dict(true=True, false=False, none=None)
 
 
 class Specificity(namedtuple('Specificity',
@@ -979,7 +997,7 @@ class StyleLog(object):
         self.entries = []
 
     def log_styled(self, styled, container, continued, custom_message=None):
-        matches = self.stylesheet.find_matches(styled)
+        matches = self.stylesheet.find_matches(styled, container.document)
         log_entry = StyleLogEntry(styled, container, matches, continued,
                                   custom_message)
         self.entries.append(log_entry)