+++ CHANGES.rst
@@ -6,10 +6,17 @@ Release 0.4.3.dev
 
 New Features:
 
+* Table column widths can be specified in the style sheet, which take effect
+  when these haven't been specified in the source document.
+* The ``page_break`` style attribute is no longer reserved for sections; a
+  page break can be forced before any flowable.
 * It's now possible to add arbitrary sections to the front/back matter by
   adding a container with the 'supporting-matter' class and a name to reference
   it by in the document template configuration, e.g. in the list of front
   matter flowables (to be documented).
+* Selectors in style sheet files (.rts) now support boolean and 'None' values.
+  For example, you can select StaticGroupedFlowables based on whether they have
+  any children or not: e.g ``TableCell(empty=true)`` selects empty table cells.
 
 Fixed: